+++ src/deps.js
@@ -1,0 +1,1 @@
+export { parse } from "https://deno.land/std@0.133.0/flags/mod.ts";
